moved detector (dis)assembly functions to psana_interface
